+++ src/commands/taskkill.js
@@ -5,7 +5,7 @@ const _fatalError = require("../functions/fatalError");
 
 const taskkill = (pid) => {
   if (typeof pid === "undefined") {
-    _errorInterpret(27);
+    _errorInterpret(2, { type: "a PID", example: "taskkill 1234" });
     return;
   }
 
@@ -15,7 +15,12 @@ const taskkill = (pid) => {
   };
 
   if (!isNumeric(pid)) {
-    _errorInterpret(28);
+    _errorInterpret(10, {
+      type: "PID",
+      supposedTo: "numbers",
+      notContain: "letters/symbols",
+      variable: pid,
+    });
     return;
   }
 
@@ -26,7 +31,7 @@ const taskkill = (pid) => {
     console.log(`Successfully killed process ${chalk.green(pid)}.\n`);
   } catch (err) {
     if (err.code === "EPERM") {
-      _errorInterpret(29, { variable: pid });
+      _errorInterpret(4, { todo: "kill the process with PID", variable: pid });
     } else {
       _fatalError(err);
     }